requirements.txt
python-telegram-bot[rate-limiter]==20.1
tiktoken>=0.3.0
# Note that PyYAML==6.0 breaks with Python3.12
PyYAML==6.0.1
pymongo==4.3.3
python-dotenv==0.21.0
pydub==0.25.1
azure.cognitiveservices.speech
uuid
requests
chardet
llama-hub>=0.0.79.post1
llama-index-llms-azure-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-index-embeddings-azure-openai
llama-index-agent-openai
llama-index-tools-bing-search
llama-index-tools-weather
llama-index-tools-google
llama-index-tools-wikipedia
pyowm
llama-index
langchain>=0.0.329
youtube_transcript_api
supabase
docx2txt
pypdf
cohere
groq
google.generativeai
pinecone
lxml_html_clean
soundfile
sounddevice
gpt_researcher
